CLOUDP-295785 - ass skip_if_exists flag to all releasable images
