+++ build_info.json
@@ -21,6 +21,7 @@
       "release": {
         "sign": true,
         "olm-tag": true,
+        "skip-if-exists": true,
         "repositories": ["quay.io/mongodb/mongodb-kubernetes"],
         "platforms": [
           "linux/arm64",
@@ -68,6 +69,7 @@
       "release": {
         "sign": true,
         "olm-tag": true,
+        "skip-if-exists": true,
         "repositories": ["quay.io/mongodb/mongodb-kubernetes-init-database"],
         "platforms": [
           "linux/arm64",
@@ -99,6 +101,7 @@
       "release": {
         "sign": true,
         "olm-tag": true,
+        "skip-if-exists": true,
         "repositories": ["quay.io/mongodb/mongodb-kubernetes-init-appdb"],
         "platforms": [
           "linux/arm64",
@@ -127,6 +130,7 @@
       "release": {
         "sign": true,
         "olm-tag": true,
+        "skip-if-exists": true,
         "repositories": ["quay.io/mongodb/mongodb-kubernetes-init-ops-manager"],
         "platforms": [
           "linux/amd64"
@@ -155,6 +159,7 @@
       "release": {
         "sign": true,
         "olm-tag": true,
+        "skip-if-exists": true,
         "repositories": ["quay.io/mongodb/mongodb-kubernetes-database"],
         "platforms": [
           "linux/arm64",
@@ -180,6 +185,7 @@
         ]
       },
       "release": {
+        "skip-if-exists": true,
         "repositories": ["quay.io/mongodb/mongodb-kubernetes-tests"],
         "platforms": [
           "linux/amd64"
@@ -280,6 +286,7 @@
       "manual_release": {
         "sign": true,
         "olm-tag": true,
+        "skip-if-exists": true,
         "repositories": ["quay.io/mongodb/mongodb-agent-ubi", "quay.io/mongodb/mongodb-agent"],
         "platforms": [
           "linux/arm64",
@@ -307,6 +314,7 @@
       "manual_release": {
         "sign": true,
         "olm-tag": true,
+        "skip-if-exists": true,
         "repositories": ["quay.io/mongodb/mongodb-enterprise-ops-manager-ubi"],
         "platforms": [
           "linux/amd64"
