+++ scripts/release/check_changelog.py
@@ -3,20 +3,9 @@ import pathlib
 
 from git import Repo
 
+from scripts.release.argparse_utils import str2bool
 from scripts.release.changelog import get_changelog_entries
 from scripts.release.constants import DEFAULT_CHANGELOG_PATH, DEFAULT_REPOSITORY_PATH
-
-
-def str2bool(v):
-    if isinstance(v, bool):
-        return v
-    if v.lower() in ("yes", "true", "t", "y", "1"):
-        return True
-    elif v.lower() in ("no", "false", "f", "n", "0"):
-        return False
-    else:
-        raise argparse.ArgumentTypeError("Boolean value expected.")
-
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(
