+++ scripts/release/tests/build_info_test.py
@@ -290,6 +290,7 @@ def test_load_build_info_release():
                 repositories=["quay.io/mongodb/mongodb-kubernetes"],
                 platforms=["linux/arm64", "linux/amd64", "linux/s390x", "linux/ppc64le"],
                 dockerfile_path="docker/mongodb-kubernetes-operator/Dockerfile",
+                skip_if_exists=True,
                 olm_tag=True,
                 sign=True,
             ),
@@ -297,6 +298,7 @@ def test_load_build_info_release():
                 repositories=["quay.io/mongodb/mongodb-kubernetes-init-database"],
                 platforms=["linux/arm64", "linux/amd64", "linux/s390x", "linux/ppc64le"],
                 dockerfile_path="docker/mongodb-kubernetes-init-database/Dockerfile",
+                skip_if_exists=True,
                 olm_tag=True,
                 sign=True,
             ),
@@ -304,6 +306,7 @@ def test_load_build_info_release():
                 repositories=["quay.io/mongodb/mongodb-kubernetes-init-appdb"],
                 platforms=["linux/arm64", "linux/amd64", "linux/s390x", "linux/ppc64le"],
                 dockerfile_path="docker/mongodb-kubernetes-init-appdb/Dockerfile",
+                skip_if_exists=True,
                 olm_tag=True,
                 sign=True,
             ),
@@ -311,6 +314,7 @@ def test_load_build_info_release():
                 repositories=["quay.io/mongodb/mongodb-kubernetes-init-ops-manager"],
                 platforms=["linux/amd64"],
                 dockerfile_path="docker/mongodb-kubernetes-init-ops-manager/Dockerfile",
+                skip_if_exists=True,
                 olm_tag=True,
                 sign=True,
             ),
@@ -318,6 +322,7 @@ def test_load_build_info_release():
                 repositories=["quay.io/mongodb/mongodb-kubernetes-database"],
                 platforms=["linux/arm64", "linux/amd64", "linux/s390x", "linux/ppc64le"],
                 dockerfile_path="docker/mongodb-kubernetes-database/Dockerfile",
+                skip_if_exists=True,
                 olm_tag=True,
                 sign=True,
             ),
@@ -325,6 +330,7 @@ def test_load_build_info_release():
                 repositories=["quay.io/mongodb/mongodb-kubernetes-tests"],
                 platforms=["linux/amd64"],
                 dockerfile_path="docker/mongodb-kubernetes-tests/Dockerfile",
+                skip_if_exists=True,
             ),
             "readiness-probe": ImageInfo(
                 repositories=["quay.io/mongodb/mongodb-kubernetes-readinessprobe"],
@@ -370,6 +376,7 @@ def test_load_build_info_manual_release():
                 repositories=["quay.io/mongodb/mongodb-agent-ubi", "quay.io/mongodb/mongodb-agent"],
                 platforms=["linux/arm64", "linux/amd64", "linux/s390x", "linux/ppc64le"],
                 dockerfile_path="docker/mongodb-agent/Dockerfile",
+                skip_if_exists=True,
                 olm_tag=True,
                 sign=True,
             ),
@@ -377,6 +384,7 @@ def test_load_build_info_manual_release():
                 repositories=["quay.io/mongodb/mongodb-enterprise-ops-manager-ubi"],
                 platforms=["linux/amd64"],
                 dockerfile_path="docker/mongodb-enterprise-ops-manager/Dockerfile",
+                skip_if_exists=True,
                 olm_tag=True,
                 sign=True,
             ),
